minor syntactic fix (#10)
(What follows is an observation that is hoped to be helpful, though it may not be an iron-clad rule, since English is a hodgepodge of many languages, and therefore not very consistent.  Also, it's just an off-the-cuff observation, rather than one I read in a grammar book that has been tested by generations of grammarians.  Nonetheless...)  "than" is used for contrasts, like "a Crystal app takes less CPU power than a Ruby app", or "Ruby is cooler than Java", etc., but "as" goes in phrases following the words "the same", and often other phrases where similarity is expressed.  For instance, "Crystal syntax is as easy to learn *as* Ruby syntax, but the executables run much faster *than* Ruby scripts."